requirements.txt
numpy
langchain
langchain_core
langchain_groq
langchain_community
pymongo
fastapi
uvicorn
python-multipart
pydantic
python-dotenv